Am adaugat partea de REACT frontend, am creat routes, folderul components ca sa comunice elementele din model si controller din backend sa ajunga la frontend

+++ src/main/java/ro/licenta/taberemanager/model/ParticipantPrezenta.java
@@ -21,9 +21,9 @@ private Long id;
     private String observatii;
 @NotNull
 @Column(name="id_activitate")
-    private Long idActivitate;
+    private BigDecimal idActivitate;
 @NotNull
     @Column(name="id_participant")
-    private Long idParticipant;
+    private BigDecimal idParticipant;
 
 }
