+++ src/main/java/ro/licenta/taberemanager/model/TabaraCategorie.java
@@ -3,6 +3,8 @@ import jakarta.persistence.*;
 import  lombok.Data;
 import lombok.NoArgsConstructor;
 import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Date;
 
@@ -17,10 +19,10 @@ public class TabaraCategorie {
     private Long id;
     @NotNull
     @Column(name="id_tabara")
-    private Long idTabara;
+    private BigDecimal idTabara;
     @NotNull
     @Column(name="id_categorie")
-    private Long idCategorie;
+    private BigDecimal idCategorie;
 
 
 }
